- DEVSU-2312 - Update appendix type, template type, and project type - Fix bug where useEffect was called on an infinite loop to reflect changes to appendices after adding new one by updating useCallback and onClose data when add appendix dialog is closed - Reorganize several data type declarations - Code linting
