+++ app/common.d.ts
@@ -86,13 +86,21 @@ type UserType = {
   allowNotifications: boolean;
 } & RecordDefaults;
 
+type ShortReportType = {
+  alternateIdentifier: string | null;
+  patientId: string;
+} & RecordDefaults;
+
 type ProjectType = {
   name: string;
+  description: string;
+  reports: ShortReportType[];
+  users: UserType[];
 } & RecordDefaults;
 
 type AppendixType = {
-  template: string;
-  project: string;
+  template: TemplateType;
+  project: ProjectType;
   text: string;
 } & RecordDefaults;
 
@@ -343,6 +351,8 @@ type MicrobialType = {
 
 export {
   RecordDefaults,
+  ShortReportType,
+  ProjectType,
   AppendixType,
   UserType,
   TemplateType,
